requirements.txt
requests==2.31.0
customtkinter==5.2.2
Pillow==10.0.1
pyyaml==6.0.1
schedule==1.2.0
